Move eere messages out of EEREChart component and into their own EEREMessages component
